refactor : 헤드/네브바 분리, user내에 common폴더에서 헤더,네브,푸터 관리
