+++ src/App.jsx
@@ -1,17 +1,18 @@
 import { BrowserRouter, Routes, Route, Navigate } from 'react-router';
 import AdminPage from './admin/pages/AdminPage';
 import UserPage from './user/pages/UserPage';
-import LoginPage from './user/components/LoginPage';
+
 import './App.css';
 
 function App() {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<Navigate to="/user" replace />} />
+        <Route path="/" element={<UserPage />} />
         <Route path="/admin" element={<AdminPage />} />
         <Route path="/user" element={<UserPage />} />
-        <Route path="/login" element={<LoginPage />} />
+
+        
       </Routes>
     </BrowserRouter>
   );
